Support parsing standalone match criteria
ssh_config(5) defines three Match criteria that don't take any
arguments: 'all', 'canonical', and 'final'. Update the criteria parsing
logic to recognise these.

The value is irrelevant, but the criteria are stored in a map, so an
empty array is chosen here. Consumers really shouldn't be trying to
use any particular value though.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>