Improve support for obscure ssh_config(5) rules (#68)
* Better handling for trailing blank lines

It's reasonable to end the config with a section directive (e.g.,
Match final' to trigger a re-parse), in which case there is no prior
line within the config to attach the whitespace to. Extend the logic
to annotate the most recent section directive if it's sub-config
is empty.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

* Support parsing standalone match criteria

ssh_config(5) defines three Match criteria that don't take any
arguments: 'all', 'canonical', and 'final'. Update the criteria parsing
logic to recognise these.

The value is irrelevant, but the criteria are stored in a map, so an
empty array is chosen here. Consumers really shouldn't be trying to
use any particular value though.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

* Test standalone match criteria parsing

Add some testing to assert that Match criteria without arguments ('all',
'canonical', and 'final') are parsed correctly.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

* Set User field if passed on the command line

When User is passed in the params, it acts as if the config started with
an unconditional User line.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

* Test .compute explicit user takes priority

When a user is passed explicitly to .compute, it should not switch
because of any User lines.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

* Extend .compute to handle Match all/final

Computing a config potentially performs multiple passes, with some
quirks around what exactly the host is in 'Host' and 'Match host'
sections.

Extend .compute to handle more cases of 'all' and 'final'.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

* Test .compute with standalone Match criteria

Test some cases of parsing configs with second passes due to Match final
or similar.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

---------

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>