Export public interface
Exports the types involved in the public interface to remain compatible
with existing users.

The object extension of the SSHConfig class is converted to static
methods to allow SSHConfig to be used as a type.

Separator is exported because it is the type of Directive::separator,
but Space is only used internally so not exported.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>